+++ crates/ui2/Cargo.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
+[lib]
+name = "ui2"
+path = "src/ui2.rs"
+
 [dependencies]
 anyhow.workspace = true
 chrono = "0.4"
