+++ crates/ui2/src/components.rs
@@ -2,56 +2,32 @@ mod avatar;
 mod button;
 mod checkbox;
 mod context_menu;
-mod details;
 mod divider;
-mod elevated_surface;
-mod facepile;
 mod icon;
 mod icon_button;
-mod indicator;
 mod input;
 mod keybinding;
 mod label;
 mod list;
-mod modal;
-mod notification_toast;
-mod palette;
-mod panel;
-mod player;
-mod player_stack;
 mod slot;
 mod stack;
-mod tab;
-mod toast;
+mod stories;
 mod toggle;
-mod tool_divider;
 mod tooltip;
 
 pub use avatar::*;
 pub use button::*;
 pub use checkbox::*;
 pub use context_menu::*;
-pub use details::*;
 pub use divider::*;
-pub use elevated_surface::*;
-pub use facepile::*;
 pub use icon::*;
 pub use icon_button::*;
-pub use indicator::*;
 pub use input::*;
 pub use keybinding::*;
 pub use label::*;
 pub use list::*;
-pub use modal::*;
-pub use notification_toast::*;
-pub use palette::*;
-pub use panel::*;
-pub use player::*;
-pub use player_stack::*;
 pub use slot::*;
 pub use stack::*;
-pub use tab::*;
-pub use toast::*;
+pub use stories::*;
 pub use toggle::*;
-pub use tool_divider::*;
 pub use tooltip::*;
