--- crates/ui2/src/components/details.rs
@@ -1,83 +1,0 @@
-use crate::prelude::*;
-use crate::{v_stack, ButtonGroup};
-
-#[derive(RenderOnce)]
-pub struct Details {
-    text: &'static str,
-    meta: Option<&'static str>,
-    actions: Option<ButtonGroup>,
-}
-
-impl Component for Details {
-    type Rendered = Div;
-
-    fn render(self, cx: &mut WindowContext) -> Self::Rendered {
-        v_stack()
-            .p_1()
-            .gap_0p5()
-            .text_ui_sm()
-            .text_color(cx.theme().colors().text)
-            .size_full()
-            .child(self.text)
-            .children(self.meta.map(|m| m))
-            .children(self.actions.map(|a| a))
-    }
-}
-
-impl Details {
-    pub fn new(text: &'static str) -> Self {
-        Self {
-            text,
-            meta: None,
-            actions: None,
-        }
-    }
-
-    pub fn meta_text(mut self, meta: &'static str) -> Self {
-        self.meta = Some(meta);
-        self
-    }
-
-    pub fn actions(mut self, actions: ButtonGroup) -> Self {
-        self.actions = Some(actions);
-        self
-    }
-}
-
-use gpui::{Div, RenderOnce};
-#[cfg(feature = "stories")]
-pub use stories::*;
-
-#[cfg(feature = "stories")]
-mod stories {
-    use super::*;
-    use crate::{Button, Story};
-    use gpui::{Div, Render};
-
-    pub struct DetailsStory;
-
-    impl Render for DetailsStory {
-        type Element = Div;
-
-        fn render(&mut self, cx: &mut ViewContext<Self>) -> Self::Element {
-            Story::container(cx)
-                .child(Story::title_for::<Details>(cx))
-                .child(Story::label(cx, "Default"))
-                .child(Details::new("The quick brown fox jumps over the lazy dog"))
-                .child(Story::label(cx, "With meta"))
-                .child(
-                    Details::new("The quick brown fox jumps over the lazy dog")
-                        .meta_text("Sphinx of black quartz, judge my vow."),
-                )
-                .child(Story::label(cx, "With meta and actions"))
-                .child(
-                    Details::new("The quick brown fox jumps over the lazy dog")
-                        .meta_text("Sphinx of black quartz, judge my vow.")
-                        .actions(ButtonGroup::new(vec![
-                            Button::new("Decline"),
-                            Button::new("Accept").variant(crate::ButtonVariant::Filled),
-                        ])),
-                )
-        }
-    }
-}
