--- crates/ui2/src/components/modal.rs
@@ -1,85 +1,0 @@
-use gpui::{AnyElement, Div, RenderOnce};
-use smallvec::SmallVec;
-
-use crate::{h_stack, prelude::*, v_stack, Button, Icon, IconButton, Label};
-
-#[derive(RenderOnce)]
-pub struct Modal {
-    id: ElementId,
-    title: Option<SharedString>,
-    primary_action: Option<Button>,
-    secondary_action: Option<Button>,
-    children: SmallVec<[AnyElement; 2]>,
-}
-
-impl Component for Modal {
-    type Rendered = gpui::Stateful<Div>;
-
-    fn render(self, cx: &mut WindowContext) -> Self::Rendered {
-        v_stack()
-            .id(self.id.clone())
-            .w_96()
-            // .rounded_xl()
-            .bg(cx.theme().colors().background)
-            .border()
-            .border_color(cx.theme().colors().border)
-            .shadow_2xl()
-            .child(
-                h_stack()
-                    .justify_between()
-                    .p_1()
-                    .border_b()
-                    .border_color(cx.theme().colors().border)
-                    .child(div().children(self.title.clone().map(|t| Label::new(t))))
-                    .child(IconButton::new("close", Icon::Close)),
-            )
-            .child(v_stack().p_1().children(self.children))
-            .when(
-                self.primary_action.is_some() || self.secondary_action.is_some(),
-                |this| {
-                    this.child(
-                        h_stack()
-                            .border_t()
-                            .border_color(cx.theme().colors().border)
-                            .p_1()
-                            .justify_end()
-                            .children(self.secondary_action)
-                            .children(self.primary_action),
-                    )
-                },
-            )
-    }
-}
-
-impl Modal {
-    pub fn new(id: impl Into<ElementId>) -> Self {
-        Self {
-            id: id.into(),
-            title: None,
-            primary_action: None,
-            secondary_action: None,
-            children: SmallVec::new(),
-        }
-    }
-
-    pub fn title(mut self, title: impl Into<SharedString>) -> Self {
-        self.title = Some(title.into());
-        self
-    }
-
-    pub fn primary_action(mut self, action: Button) -> Self {
-        self.primary_action = Some(action);
-        self
-    }
-
-    pub fn secondary_action(mut self, action: Button) -> Self {
-        self.secondary_action = Some(action);
-        self
-    }
-}
-
-impl ParentElement for Modal {
-    fn children_mut(&mut self) -> &mut SmallVec<[AnyElement; 2]> {
-        &mut self.children
-    }
-}
