+++ crates/ui2/src/components/stories/button.rs
@@ -1,0 +1,167 @@
+#[cfg(feature = "stories")]
+pub use stories::*;
+
+#[cfg(feature = "stories")]
+mod stories {
+    use super::*;
+    use crate::{h_stack, v_stack, Story, TextColor};
+    use gpui::{rems, Div, Render};
+    use strum::IntoEnumIterator;
+
+    pub struct ButtonStory;
+
+    impl Render for ButtonStory {
+        type Element = Div;
+
+        fn render(&mut self, cx: &mut ViewContext<Self>) -> Self::Element {
+            let states = InteractionState::iter();
+
+            Story::container(cx)
+                .child(Story::title_for::<Button>(cx))
+                .child(
+                    div()
+                        .flex()
+                        .gap_8()
+                        .child(
+                            div()
+                                .child(Story::label(cx, "Ghost (Default)"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label").variant(ButtonVariant::Ghost), // .state(state),
+                                        )
+                                })))
+                                .child(Story::label(cx, "Ghost – Left Icon"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label")
+                                                .variant(ButtonVariant::Ghost)
+                                                .icon(Icon::Plus)
+                                                .icon_position(IconPosition::Left), // .state(state),
+                                        )
+                                })))
+                                .child(Story::label(cx, "Ghost – Right Icon"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label")
+                                                .variant(ButtonVariant::Ghost)
+                                                .icon(Icon::Plus)
+                                                .icon_position(IconPosition::Right), // .state(state),
+                                        )
+                                }))),
+                        )
+                        .child(
+                            div()
+                                .child(Story::label(cx, "Filled"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label").variant(ButtonVariant::Filled), // .state(state),
+                                        )
+                                })))
+                                .child(Story::label(cx, "Filled – Left Button"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label")
+                                                .variant(ButtonVariant::Filled)
+                                                .icon(Icon::Plus)
+                                                .icon_position(IconPosition::Left), // .state(state),
+                                        )
+                                })))
+                                .child(Story::label(cx, "Filled – Right Button"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label")
+                                                .variant(ButtonVariant::Filled)
+                                                .icon(Icon::Plus)
+                                                .icon_position(IconPosition::Right), // .state(state),
+                                        )
+                                }))),
+                        )
+                        .child(
+                            div()
+                                .child(Story::label(cx, "Fixed With"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label")
+                                                .variant(ButtonVariant::Filled)
+                                                // .state(state)
+                                                .width(Some(rems(6.).into())),
+                                        )
+                                })))
+                                .child(Story::label(cx, "Fixed With – Left Icon"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label")
+                                                .variant(ButtonVariant::Filled)
+                                                // .state(state)
+                                                .icon(Icon::Plus)
+                                                .icon_position(IconPosition::Left)
+                                                .width(Some(rems(6.).into())),
+                                        )
+                                })))
+                                .child(Story::label(cx, "Fixed With – Right Icon"))
+                                .child(h_stack().gap_2().children(states.clone().map(|state| {
+                                    v_stack()
+                                        .gap_1()
+                                        .child(
+                                            Label::new(state.to_string()).color(TextColor::Muted),
+                                        )
+                                        .child(
+                                            Button::new("Label")
+                                                .variant(ButtonVariant::Filled)
+                                                // .state(state)
+                                                .icon(Icon::Plus)
+                                                .icon_position(IconPosition::Right)
+                                                .width(Some(rems(6.).into())),
+                                        )
+                                }))),
+                        ),
+                )
+                .child(Story::label(cx, "Button with `on_click`"))
+                .child(
+                    Button::new("Label")
+                        .variant(ButtonVariant::Ghost)
+                        .on_click(|_, cx| println!("Button clicked.")),
+                )
+        }
+    }
+}
