+++ crates/ui2/src/components/stories/mod.rs
@@ -1,0 +1,8 @@
+mod avatar;
+mod button;
+mod checkbox;
+mod context_menu;
+mod icon;
+mod input;
+mod keybinding;
+mod label;
