+++ crates/ui2/src/ui2.rs
@@ -15,15 +15,15 @@
 #![allow(dead_code, unused_variables)]
 
 mod components;
-mod elevation;
 pub mod prelude;
 mod styled_ext;
+mod styles;
 pub mod utils;
 
 pub use components::*;
 pub use prelude::*;
-// pub use static_data::*;
 pub use styled_ext::*;
+pub use styles::*;
 
 #[cfg(feature = "stories")]
 mod story;
